Updated scripts for generating MSL patches
The new scripts use OMC API to update annotations, etc. This should
be more resistant to changes in MSL. Also bumped MSL version.

Note that the new script is not used by the Hudson build yet. It will
be added in the future (once nightly builds have the required
capability).